Update styles.css
Congratulations on the application, very simple and functional. I had trouble finding the "CEP" address mentioned, as I'm not from the country of origin of this project, but I looked for some random ones on the internet and it worked. I just made a change to the colors that I thought were too exaggerated and unconnected, otherwise it's great, I'll even get the idea.